ImportPathExcludesCommand: Allow for reusing importPathExcludes()
